requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
pydantic-settings==2.5.2

# LLM & AI
openai==1.51.0

# RAG & Vector DB
faiss-cpu>=1.7.4
numpy>=1.24.0

# Utils
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12

# Dev
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.8.0
